chore(notification-dismiss-button): move to common, rename to CloseButton
