+++ packages/components/src/common/CloseButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CloseButton";
