--- packages/components/src/common/Notifications/NotificationDismissButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./NotificationDismissButton";
